Changed Discount amount to float field

+++ final_project/e_drugs/models.py
@@ -9,7 +9,7 @@ class Alert(Model):
 
 class Discount(Model):
     name = CharField(max_length=128)
-    amount = IntegerField()
+    amount = FloatField()
 
 
 class Shipping(Model):
